fix(new): handle --owner and --standalone flags in JSON mode
Previously, `bwrb new --json` silently ignored ownership flags, always
creating pooled notes. This violated the principle of predictable CLI
behavior and broke automation use cases.

Changes:
- Pass ownership options to createNoteFromJson()
- Validate mutual exclusivity of --owner and --standalone
- Create owned notes when --owner is provided
- Return clear JSON errors for invalid ownership scenarios
- Error when --standalone used on non-ownable type
- Fix owner arg parsing order (quotes first, then brackets)

Test changes:
- Add new-ownership.test.ts with isolated vault for ownership tests
- Keep ownership types separate to avoid affecting other tests

Fixes #138
